docs: add comprehensive SIMD optimization and bit-level pattern guides
Add four new skill documents documenting advanced optimization patterns
discovered during development:

- Add bit-optimization.md covering byte-level lookup tables, cumulative
  indices, and hierarchical skip patterns with concrete examples
- Add simd-patterns.md documenting SIMD compilation model, instruction
  set hierarchy, testing strategies, and multi-architecture patterns
- Update json-semi-indexing.md with performance optimization section
  showing 627x speedup from O(n) to O(log n) select operations
- Update rust-no-std-simd.md with bit position conventions, compile-time
  lookup tables, and zero-copy output patterns
- Update CLAUDE.md with restructured architecture section and clearer
  documentation organization
- Fix minor type casting in bp.rs find_close_in_word_fast

These documents capture institutional knowledge about when and how to
apply SIMD optimizations, avoid common pitfalls, and achieve significant
performance gains through proper algorithm selection.

+++ src/bp.rs
@@ -153,7 +153,7 @@ fn find_close_in_word_fast(
     if bit_in_byte != 0 {
         // Scan remaining bits in first partial byte
         let byte_val = bytes[first_byte_idx];
-        let end_bit = (8.min(valid_bits - first_byte_idx * 8)) as usize;
+        let end_bit = 8.min(valid_bits - first_byte_idx * 8);
 
         for bit in bit_in_byte..end_bit {
             if (byte_val >> bit) & 1 == 1 {
